backend/requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
gunicorn==23.0.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
python-whois==0.9.5
requests==2.32.3
scikit-learn==1.6.1
